refactored the current code.
